backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# Monitoring dependencies
pynput==1.7.6
psutil==5.9.8
pillow==10.2.0
pygetwindow==0.0.9
pystray==0.19.5

# API utilities
pydantic==2.5.3
pydantic-settings==2.1.0

# Data export
reportlab==4.0.9
pandas==2.2.0

# WebSocket support
websockets==12.0

# Development tools
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Executable creation
pyinstaller==6.3.0
